+++ Source/Tests/DownloadTestCase.swift
@@ -17,8 +17,6 @@ class DownloadTestCase: ProtocolStubbedTestCase {
     let searchPathDomain = FileManager.SearchPathDomainMask.userDomainMask
 
     func testDownloadRequest() {
-        // Given
-
         let destination = Alamofire.DownloadRequest.suggestedDownloadDestination(
             for: searchPathDirectory,
             in: searchPathDomain
@@ -57,4 +55,47 @@ class DownloadTestCase: ProtocolStubbedTestCase {
         waitForExpectations(timeout: 1, handler: nil)
         XCTAssertEqual(resultingRequest.task?.currentRequest?.timeoutInterval, 3)
     }
+    
+    func testDownloadSuccessBlock() {
+        let destination = Alamofire.DownloadRequest.suggestedDownloadDestination(
+            for: searchPathDirectory,
+            in: searchPathDomain
+        )
+        let responseSerializer = TRONDownloadResponseSerializer { _, _, url, _ in url }
+        let request: DownloadAPIRequest<URL?, APIError> = tron
+            .download("/stream/100",
+                      to: destination,
+                      responseSerializer: responseSerializer)
+            .stubSuccess(.init(), statusCode: 200)
+        let expectation = self.expectation(description: "Download expectation")
+        request.perform(withSuccess: { url in
+            expectation.fulfill()
+        }) { _ in
+            XCTFail()
+        }
+        waitForExpectations(timeout: 1, handler: nil)
+    }
+    
+    func testDownloadFailureBlock() {
+        let destination = Alamofire.DownloadRequest.suggestedDownloadDestination(
+            for: searchPathDirectory,
+            in: searchPathDomain
+        )
+        let responseSerializer = TRONDownloadResponseSerializer<URL?> { _, _, url, _ in throw "Fail" }
+        let request: DownloadAPIRequest<URL?, APIError> = tron
+            .download("/stream/100",
+                      to: destination,
+                      responseSerializer: responseSerializer)
+            .stubSuccess(.init(), statusCode: 200)
+        let expectation = self.expectation(description: "Download expectation")
+        request.perform(withSuccess: { url in
+            XCTFail()
+        }) { error in
+            XCTAssertEqual(error.error as? String, "Fail")
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: 1, handler: nil)
+    }
 }
+
+extension String: Error {}